Added FITC functionality
Added the FITC functionality very similar to GPML Matlab toolbox.  Had
to rewrite all the covariance and mean functions to fit the format.
Also, fixed a bug in covNoise.  The code is much slower now, and will
need to be improved.  Especially in the FITC inference.  I deviated
away from a speedy implementation for clarity.